intern_project/requirements.txt
fastapi
uvicorn
mysql-connector-python
boto3
